filter Kakarot Transactions
